Move export schemas to a separate file
